Complete empty uploads
The upload completer scans for uploads that need to be completed,
likely due to an error or process restart. Prior to this change,
it only completed uploads that had 1 or more parts. Since completing
an upload is what cleans up the directory on disk, it was possible
for us to leave behind empty directories for uploads with no parts.

This change makes it valid to complete uploads with no parts, which
ensures that these directories get cleaned up.

In addition, the upload completer checks the upload for a session.end
event, and emits one if there is not already an end event present.
This logic only handled session.end, and not
windows.desktop.session.end, so that has been updated as well.

Updates #9646